The definitive 100% broken commit!

+++ MagicKeys/Core/Structs.cs
@@ -62,7 +62,7 @@ public IntPtr dwExtraInfo;
 public struct KeyboardInputData
 {
 public int wVk;
-public ushort wScan;
+public int wScan;
 public KeyEventFlags dwFlags;
 public int time;
 public IntPtr dwExtraInfo;
